+++ tests/v19/models/order/test_data.py
@@ -1,22 +1,6 @@
 """Tests for the data module."""
 
-import pytest
-
 from csobpg.v19.models import currency, order
-
-
-@pytest.mark.parametrize(
-    "quantity",
-    [
-        0,
-        -1,
-        100,
-    ],
-)
-def test_gift_cards_invalid_quantity(quantity: int):
-    """Test invalid quantity arg for the GiftCardsData."""
-    with pytest.raises(ValueError, match="should be"):
-        order.GiftCardsData(quantity=quantity)
 
 
 def test_gift_cards_unset_fields_are_not_signed():